requirements.txt
fastapi==0.128.1
uvicorn==0.40.0
SQLAlchemy==2.0.46
python-dotenv==1.2.1
pydantic==2.12.5
google-auth==2.29.0
bcrypt==4.1.3
requests==2.32.5
email-validator==2.3.0
ruff==0.5.7
pytest==9.0.2
google-genai==1.55.0